Moved icon styles to shared folder.
